requirements.txt
# Core dependencies
kivy==2.3.0
python-dotenv==1.0.0

# Kivy dependencies (for Python 3.11)
kivy-deps.sdl2==0.7.0
kivy-deps.glew==0.3.1
kivy-deps.gstreamer==0.3.4
kivy-deps.angle==0.4.0

# Communication
requests==2.31.0
plyer==2.1.0
twilio==8.5.0

# Database
sqlalchemy==2.0.27
alembic==1.13.1
bcrypt==4.1.2
psycopg2-binary==2.9.9  # For PostgreSQL support

# Additional utilities
pillow==10.2.0  # For image handling

# Development dependencies
cython==0.29.33

# KivyMD
kivymd==1.1.1

# Additional dependencies
pyjnius==1.6.1
buildozer==1.5.0
